Centering api checkin code
