+++ api/api.ts
@@ -1,17 +1,12 @@
 import express from 'express'
 import * as contentDisposition from 'content-disposition'
 import bodyParser from 'body-parser'
-import { config as dotenvConfig } from "dotenv-safe"
-import { getTenantConfig } from '../db/db.js'
-import { initializeDriver, uploadFile, downloadFile } from '../drivers/driver_manager.js'
+import checkinDefault from './checkin.js'
+import { uploadFile, downloadFile } from '../drivers/driver_manager.js'
 
 const MAX_SHARED_API_FILE_SIZE = '5mb'
 const app = express()
 app.use(bodyParser.raw({type: 'application/octet-stream', limit : MAX_SHARED_API_FILE_SIZE}))
-
-dotenvConfig()
-
-const authUrl = process.env.URL_AUTH_SERVICE;
 
 app.post('/upload/:tenant', async (req: express.Request, res: express.Response) => {
     if (req.headers['content-type'] !== 'application/octet-stream') {
@@ -20,48 +15,15 @@ app.post('/upload/:tenant', async (req: express.Request, res: express.Response) 
         return
     }
 
-    const authHeader = req.headers['authorization']
-    if (!authHeader) {
-        console.error(`Missing Authorization header at ${new Date()}`)
-        res.status(401).send('Missing Authorization header')
-        return
-    }
+    const checkinResult = await checkinDefault(req.headers['authorization'] as string, req.params.tenant)
 
-    const token = authHeader?.split(' ')[1]
-    if (!token) {
-        console.error(`Request missing bearer token at ${new Date()}`)
-        res.status(401).send('Missing bearer token')
-        return
-    }
-
-    const authParams = new URLSearchParams({
-        token: token,
-        tenant: req.params.tenant
-    });
-
-    const authResponse = await makeAuth(authUrl, authParams)
-
-    if (authResponse.status !== 200) {
-        console.error(`Error authenticating user. Caused by:\n${ 'errorMessage' in authResponse ? authResponse.errorMessage: 'not known' }\nat ${new Date()}`)
-        res.status(authResponse.status).send('Error trying authentication.')
+    if (checkinResult != null) {
+        console.error(`${checkinResult.errorMessage} at ${new Date()}`)
+        res.status(checkinResult.status).send(checkinResult.errorMessage)
         return
     }
 
     const filename: string = getFilename(req)
-    const tenantConfig = await getTenantConfig(req.params.tenant)
-
-    if (!tenantConfig) {
-        console.error(`Tenant not found at ${new Date()}`)
-        res.status(404).send('Tenant not found.')
-        return
-    }
-
-    if (!initializeDriver(tenantConfig)) {
-        console.error(`Driver not found at ${new Date()}`)
-        res.status(404).send('Driver not found.')
-        return
-    }
-
     const uploadJsonResponse = await uploadFile(req.body, filename)
 
     if ('errorMessage' in uploadJsonResponse) {
@@ -75,43 +37,11 @@ app.post('/upload/:tenant', async (req: express.Request, res: express.Response) 
 })
 
 app.get('/download/:tenant/:idFile', async (req: express.Request, res: express.Response) => {
-    const authHeader = req.headers['authorization']
-    if (!authHeader) {
-        console.error(`Missing Authorization header at ${new Date()}`)
-        res.status(401).send('Missing Authorization header')
-        return
-    }
+    const checkinResult = await await checkinDefault(req.headers['authorization'] as string, req.params.tenant)
 
-    const token = authHeader?.split(' ')[1]
-    if (!token) {
-        console.error(`Request missing bearer token at ${new Date()}`)
-        res.status(401).send('Missing bearer token')
-        return
-    }
-
-    const authParams = new URLSearchParams({
-        token: token,
-        tenant: req.params.tenant
-    });
-
-    const authResponse = await makeAuth(authUrl, authParams)
-
-    if (authResponse.status !== 200) {
-        console.error(`Error authenticating user. Caused by:\n${ 'errorMessage' in authResponse ? authResponse.errorMessage: 'not known' }\nat ${new Date()}`)
-        res.status(authResponse.status).send('Error trying authentication.')
-        return
-    }
-
-    const tenantConfig = await getTenantConfig(req.params.tenant)
-    if (!tenantConfig) {
-        console.error(`Tenant not found at ${new Date()}`)
-        res.status(404).send('Tenant not found.')
-        return
-    }
-
-    if (!initializeDriver(tenantConfig)) {
-        console.error(`Driver not found at ${new Date()}`)
-        res.status(404).send('Driver not found.')
+    if (checkinResult != null) {
+        console.error(`${checkinResult.errorMessage} at ${new Date()}`)
+        res.status(checkinResult.status).send(checkinResult.errorMessage)
         return
     }
 
@@ -136,18 +66,6 @@ app.get('/download/:tenant/:idFile', async (req: express.Request, res: express.R
 
     res.send(buffer)
 })
-
-async function makeAuth(authUrl: string | undefined, authParams: URLSearchParams): Promise<AuthStatus | Err> {
-    try {
-      const response = await fetch(`${authUrl}?${authParams}`);
-        return {status: response.status};
-    } catch (error: any) {
-        return {
-            status: 500,
-            errorMessage: error
-        }
-    }
-}
 
 function getFilename(req: express.Request): string {
     const dispositionHeader: string | undefined = req.headers['content-disposition']
